requirements.txt
python-telegram-bot
web3
pyyaml
python-dotenv
